MASSIVE refactor
