+++ SlopperEngine/Rendering/SceneRenderer.cs
@@ -1,17 +1,18 @@
 using SlopperEngine.Graphics;
 using SlopperEngine.Graphics.ShadingLanguage;
-using SlopperEngine.SceneObjects.Rendering;
+using SlopperEngine.SceneObjects;
 using System.CodeDom.Compiler;
 using OpenTK.Mathematics;
 using SlopperEngine.Graphics.GPUResources.Textures;
 using OpenTK.Graphics.OpenGL4;
+using SlopperEngine.Core;
 
-namespace SlopperEngine.Core.SceneComponents;
+namespace SlopperEngine.Rendering;
 
 /// <summary>
-/// Abstract base class for all renderers.
+/// Special SceneObject for renderers - this allows for multithreading by handling them seperately
 /// </summary>
-public abstract class RenderHandler : SceneRenderer
+public abstract class SceneRenderer : SceneObject
 {
     /// <summary>
     /// The background color of the rendered frame, if used by inheriting classes.
@@ -22,7 +23,7 @@ public abstract class RenderHandler : SceneRenderer
     protected List<Camera> cameras = new();
     protected ShaderGlobals globals;
 
-    public RenderHandler()
+    public SceneRenderer()
     {
         globals = new();
     }
@@ -36,9 +37,15 @@ public abstract class RenderHandler : SceneRenderer
         cameras.Remove(cam);
     }
 
-    public override void InputUpdate(InputUpdateArgs input){}
+	/// <summary>
+    /// Updates user input, if applicable
+    /// </summary>
+    public abstract void InputUpdate(InputUpdateArgs input);
 
-    public override void Render(FrameUpdateArgs args)
+	/// <summary>
+    /// Renders the scene
+    /// </summary>
+    public void Render(FrameUpdateArgs args)
     {
 		_time += args.DeltaTime;
 		globals.Time = _time;
@@ -52,4 +59,7 @@ public abstract class RenderHandler : SceneRenderer
 
     public abstract void AddVertexMain(SyntaxTree scope, IndentedTextWriter writer);
     public abstract void AddFragmentMain(SyntaxTree scope, IndentedTextWriter writer);
+    
+    [OnRegister] void Register() => Scene!.CheckCachedComponents();
+    [OnUnregister] void Unregister(Scene scene) => scene.CheckCachedComponents();
 }
