+++ SlopperEngine/SceneObjects/Scene.cs
@@ -5,7 +5,7 @@ using SlopperEngine.Core.Collections;
 using SlopperEngine.Core.SceneComponents;
 using SlopperEngine.Core.SceneData;
 using SlopperEngine.Core.Serialization;
-using SlopperEngine.Graphics.Renderers;
+using SlopperEngine.Rendering;
 using SlopperEngine.SceneObjects.Serialization;
 
 namespace SlopperEngine.SceneObjects;
@@ -19,7 +19,7 @@ public sealed class Scene : SceneObject
     public readonly ChildList<SceneComponent> Components;
     public readonly ChildList<SceneRenderer> Renderers;
     public UpdateHandler? UpdateHandler {get; private set;}
-    public RenderHandler? RenderHandler {get; private set;}
+    public SceneRenderer? SceneRenderer {get; private set;}
     public PhysicsHandler? PhysicsHandler {get; private set;}
 
     [DontSerialize] FridgeDictionary<Type, ISceneDataContainer> _dataContainers = new();
@@ -77,7 +77,7 @@ public sealed class Scene : SceneObject
         foreach(var comp in Components.All)
         {
             comp.FrameUpdate(update);
-		}
+        }
         UpdateRegister();
         FinalizeQueues();
     }
@@ -103,7 +103,7 @@ public sealed class Scene : SceneObject
     {
         UpdateRegister();
         FinalizeQueues();
-        foreach(var rend in Renderers.AllOfType<RenderHandler>())
+        foreach(var rend in Renderers.AllOfType<SceneRenderer>())
             rend.Render(args);
         UpdateRegister();
         FinalizeQueues();
@@ -194,11 +194,11 @@ public sealed class Scene : SceneObject
     }
 
     /// <summary>
-    /// Sets the cached components (Scene.RenderHandler, Scene.UpdateHandler, etc).
+    /// Sets the cached components (Scene.SceneRenderer, Scene.UpdateHandler, etc).
     /// </summary>
     public void CheckCachedComponents()
     {
-        RenderHandler = Renderers.FirstOfType<RenderHandler>();
+        SceneRenderer = Renderers.FirstOfType<SceneRenderer>();
         UpdateHandler = Components.FirstOfType<UpdateHandler>();
         PhysicsHandler = Components.FirstOfType<PhysicsHandler>();
     }
